Removed custom deflate pool.
