+++ src/libraries/System.Net.WebSockets/tests/System.Net.WebSockets.Tests.csproj
@@ -10,7 +10,6 @@
     <Compile Include="WebSocketTestStream.cs" />
     <Compile Include="WebSocketDeflateOptionsTests.cs" />
     <Compile Include="WebSocketDeflateTests.cs" />
-    <Compile Include="ZLibStreamTests.cs" />
     <Compile Include="$(CommonTestPath)System\Net\Configuration.cs"
              Link="Common\System\Net\Configuration.cs" />
     <Compile Include="$(CommonTestPath)System\Net\Configuration.WebSockets.cs"
